Split game.js and GamePage.vue into files, took out the configuration and types as separate components.
